Create Read an entity relationship diagram 1
Read an entity relationship diagram
The information you need is sometimes split across multiple tables in the database.

What is the most common stackoverflow tag_type? What companies have a tag of that type?

To generate a list of such companies, you'll need to join three tables together.

Reference the entity relationship diagram as needed when determining which columns to use when joining tables.

Instructions 1/2

First, using the tag_type table, count the number of tags with each type.
Order the results to find the most common tag type.