Create Summarize the distribution of numeric values 2
Summarize the distribution of numeric values
Was 2017 a good or bad year for revenue of Fortune 500 companies? Examine how revenue changed from 2016 to 2017 by first looking at the distribution of revenues_change and then counting companies whose revenue increased.

Instructions 2/3

Repeat step 1, but this time, cast revenues_change as an integer to reduce the number of different values.