Create Completion time by category
Completion time by category
The evanston311 data includes a date_created timestamp from when each request was created and a date_completed timestamp for when it was completed. The difference between these tells us how long a request was open.

Which category of Evanston 311 requests takes the longest to complete?

Instructions

Compute the average difference between the completion timestamp and the creation timestamp by category.
Order the results with the largest average time to complete the request first.